customer_churn_prediction/requirements.txt
scikit-learn
xgboost
pandas
google-cloud-bigquery
openpyxl
